refactor: move the content in BridgeWebView to BridgeWebViewClient
`jsBridge` is provided by BridgeWebViewClient now
