+++ library/src/main/java/com/github/lzyzsd/jsbridge/CallBackFunction.java
@@ -2,6 +2,6 @@ package com.github.lzyzsd.jsbridge;
 
 public interface CallBackFunction {
 	
-	public void onCallBack(String data);
+	void onCallBack(String data);
 
 }
